stress-fallocate: Fix off_t printf warnings on 32 bit GNU HURD

+++ stress-fallocate.c
@@ -110,8 +110,8 @@ int stress_fallocate(
 			if (fstat(fd, &buf) < 0)
 				pr_fail(stderr, "%s: fstat on file failed", name);
 			else if (buf.st_size != opt_fallocate_bytes)
-				pr_fail(stderr, "%s: file size %zd does not match size the expected file size of %zd\n",
-					name, buf.st_size, opt_fallocate_bytes);
+				pr_fail(stderr, "%s: file size %jd does not match size the expected file size of %jd\n",
+					name, (intmax_t)buf.st_size, (intmax_t)opt_fallocate_bytes);
 		}
 
 		if (ftruncate(fd, 0) < 0)
@@ -126,8 +126,8 @@ int stress_fallocate(
 			if (fstat(fd, &buf) < 0)
 				pr_fail(stderr, "%s: fstat on file failed", name);
 			else if (buf.st_size != (off_t)0)
-				pr_fail(stderr, "%s: file size %zd does not match size the expected file size of 0\n",
-					name, buf.st_size);
+				pr_fail(stderr, "%s: file size %jd does not match size the expected file size of 0\n",
+					name, (intmax_t)buf.st_size);
 
 		}
